Abstract out rl driver
